refactor: reorganize Layout components
