+++ components/Layout/index.tsx
@@ -1,0 +1,2 @@
+export { default as HomeLayout } from "./HomeLayout";
+export { default as Layout } from "./Layout";
